requirements.txt
streamlit==1.28.1
openai==1.3.8
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
